+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Multiple-image Network Graphics (MNG)
 - Portable FloatMap (PFM)
 - Quite OK Image (QOI)
+- SHOUTcast Playlist (PLS)
 - Silicon Graphics Image (SGI)
 - Ultimate Soundtracker Module (MOD)
 - X PixMap (XPM)